move bar creation to extra class
